@@ -3,35 +3,27 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "react-native start",
+    "start": "node node_modules/react-native/local-cli/cli.js start",
     "test": "jest"
   },
   "dependencies": {
-    "lodash": "^4.17.2",
-    "moment": "^2.17.1",
-    "react": "15.4.1",
-    "react-native": "^0.41.2",
+    "react": "16.0.0-alpha.6",
+    "react-native": "0.43.4",
     "react-native-db-models": "^0.1.3",
     "react-native-side-menu": "^0.20.1",
-    "react-redux": "^4.4.6",
-    "redux": "^3.6.0",
-    "redux-logger": "^2.7.4",
-    "redux-thunk": "^2.1.0"
-  },
-  "jest": {
-    "preset": "jest-react-native"
+    "react-redux": "^5.0.4",
+    "redux": "^3.6.0"
   },
   "devDependencies": {
-    "babel-jest": "16.0.0",
-    "babel-preset-react-native": "1.9.0",
-    "eslint": "^3.17.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
-    "jest": "16.0.1",
-    "jest-react-native": "16.0.0",
-    "react-test-renderer": "15.3.2",
-    "redux-devtools": "^3.3.1"
+    "babel-jest": "19.0.0",
+    "babel-preset-react-native": "1.9.1",
+    "jest": "19.0.2",
+    "react-test-renderer": "16.0.0-alpha.6",
+    "redux-devtools": "^3.3.2",
+    "redux-logger": "^3.0.1",
+    "redux-thunk": "^2.2.0"
+  },
+  "jest": {
+    "preset": "react-native"
   }
 }
