@@ -2,8 +2,6 @@
 
 import React from 'react';
 import {
-    PropTypes,
-    StyleSheet,
     Text,
     TextInput,
     View
@@ -15,26 +13,32 @@ const FormTextInput = React.createClass({
     displayName: 'FormTextInput',
 
     propTypes: {
-        value: React.PropTypes.string,
+        addPantsName: React.PropTypes.func,
         inputRef: React.PropTypes.string,
+        fieldName: React.PropTypes.string,
+        value: React.PropTypes.string,
         labelText: React.PropTypes.string,
         placeholderText: React.PropTypes.string
     },
 
     getDefaultProps() {
         return {
-            inputRef: '',
             labelText: 'Field Name',
             placeholderText: ''
         };
     },
 
-    onEndEditing() {
-        console.log('text changed');
+    componentWillMount() {
+        // this.setState({texty: ''})
+    },
+
+    onChangeText(text) {
+        const dbFieldName = this.props.fieldName;
+        this.props.addPantsName(text);
     },
 
     handleSubmitEditing() {
-        console.log('submitted')
+        console.log('submitted');
     },
 
     render() {
@@ -42,14 +46,14 @@ const FormTextInput = React.createClass({
             <View style={FormStyles.formFieldWrapper}>
                 <Text style={FormStyles.formLabel}>{this.props.labelText}</Text>
                 <TextInput
-                    autoCapitalize={'words'}
+                    autoCapitalize='words'
                     autoCorrect={false}
-                    underlineColorAndroid="transparent"
+                    underlineColorAndroid='transparent'
                     style={FormStyles.textInput}
                     placeholder={this.props.placeholderText}
-                    onEndEditing={this.onEndEditing}
-                    ref={this.props.inputRef}
+                    onChangeText={this.onChangeText}
                     value={this.props.value}
+                    ref={this.props.inputRef}
                 />
             </View>
         );
