@@ -8,7 +8,7 @@ import {
     FAILURE_PANTS_DATA,
     COMPLETELY_RESET_DATA
 } from '../actions/PantsListActions';
-import {assign, forEach, isUndefined, toLower } from 'lodash';
+import { assign, forEach, isUndefined, toLower } from 'lodash';
 
 const initialState = {
     pantsData: {}
